requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.11.0
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.3.37
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
openai==1.76.0
orjson==3.10.16
packaging==24.2
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==80.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zstandard==0.23.0
